website_analyzer/requirements.txt
# Core dependencies
aiohttp
requests
beautifulsoup4
lxml

# Browser automation
browser-use
playwright

# AI and image processing
openai
#azure-openai
pillow
opencv-python

# Report generation
jinja2
weasyprint
plotly
matplotlib

# Configuration and utilities
pyyaml
python-dotenv
click

# Async and parallel processing
asyncio-throttle
tenacity

# Development and logging
loguru
tqdm
